Vote page first adjustments
