+++ components/header/header.module.css
@@ -350,7 +350,7 @@
 
   .containerMenuWallet {      
     position: relative;
-    padding-left: 10%;    
+    padding-left: 9%;    
   }
 
   .containerNav {
