+++ components/ssVotes/ssVotes.module.css
@@ -17,13 +17,13 @@
 .tableContainer {
       
   background: 
-    linear-gradient(#0D142E 0 0) padding-box, 
+    transparent, 
     linear-gradient(to bottom, #CD74CC, #FFBD59 , #70DD88) border-box;
-  color: #313149;
-  
+    
   border: 1px solid transparent;  
   border-radius: 12px;    
 
+  color: #313149;
   width: 100%;
   display: flex;
   flex-direction: column;
@@ -32,14 +32,19 @@
 
 .actionButtons {
   padding: 10px 0;
-  min-width: 560px !important;
-  background: #11172a !important;
+  min-width: 560px !important;  
   position: fixed;
   bottom: 30px;
   left: 50%;
   transform: translatex(-50%);
-  border-radius: 12 !important;
-  border: 1px solid rgba(126, 153, 176, 0.2);
+            
+  background: 
+    transparent, 
+    linear-gradient(to bottom, #CD74CC, #FFBD59 , #70DD88) border-box !important;
+    
+  border: 1px solid transparent !important;  
+  border-radius: 12px;    
+
   z-index: 20 !important;
 }
 
@@ -49,8 +54,8 @@
 
 .actionButtonText {
   font-size: 15px !important;
-  text-transform: capitalize !important;
-  font-weight: 700 !important;
+  text-transform: capitalize !important;  
+  font-family: 'Arista';
 }
 
 .loadingCircle {
@@ -91,11 +96,13 @@
 
 
 .buttonOverrideFixed {
-  background: rgb(23, 52, 72) !important;
-  color: rgb(6, 211, 215) !important;
+  
   font-weight: 700 !important;
   min-width: 200px !important;
   max-height: 48px;
+
+  background-image: linear-gradient(to right, #CD74CC, #FFBD59, #70DD88, #FFBD59) !important;  
+  color: #0D142E !important;    
   
 }
 
@@ -141,7 +148,7 @@
   padding: 0 20px;
       
   background: 
-    linear-gradient(#0D142E 0 0) padding-box, 
+    transparent, 
     linear-gradient(to bottom, #CD74CC, #FFBD59 , #70DD88) border-box;
   
   border: 1px solid transparent;
